requirements.txt
bentoml
sentence-transformers
boto3
python-dotenv
pinecone